performance-monitor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
aiohttp==3.9.1
python-multipart==0.0.6
jinja2==3.1.2
prometheus-client==0.19.0
psutil==5.9.6
faker==20.1.0
asyncpg==0.29.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
